Comments, Implementation, Fixes

+++ src/main/java/com/davidbracewell/apollo/linalg/DenseDoubleMatrix.java
@@ -251,7 +251,7 @@ public class DenseDoubleMatrix implements Matrix, Serializable {
 
    @Override
    public ElementType getElementType() {
-      return ElementType.FLOAT;
+      return ElementType.DOUBLE;
    }
 
    @Override
